+++ php/api/properties/details.php
@@ -65,6 +65,10 @@ try {
     $prop['primary_image'] = $primaryImage;
     $prop['images'] = $images;
 
+    if ($currentUser['role'] !== 'super_admin') {
+        unset($prop['landing_price']);
+    }
+
     sendJsonResponse(true, "Property details retrieved.", [
         "property" => $prop
     ]);